[Refactor] Relocate to `dto.response.marker`
