+++ src/main/java/boombimapi/domain/place/query/api/dto/response/marker/ViewportClusterMarkerResponse.java
@@ -1,4 +1,4 @@
-package boombimapi.domain.place.query.api.dto.response.node;
+package boombimapi.domain.place.query.api.dto.response.marker;
 
 import boombimapi.domain.place.query.api.dto.MarkerType;
 import boombimapi.global.vo.Coordinate;
