+++ src/main/java/boombimapi/domain/place/query/api/dto/response/marker/ViewportMarkerResponse.java
@@ -1,4 +1,4 @@
-package boombimapi.domain.place.query.api.dto.response.node;
+package boombimapi.domain.place.query.api.dto.response.marker;
 
 public sealed interface ViewportMarkerResponse
     permits ViewportClusterMarkerResponse, ViewportPlaceMarkerResponse {
